\name{asynchDataTD}
\docType{data}
\alias{asynchDataTD}
\alias{TD.x}
\alias{TD.y}
\title{
Generated Asynchronous Longitudinal Data with Time-Dependent Coefficients
}
\description{
For the purposes of the package examples, the
 data set was adapted from the numerical simulations of the original
manuscript. Specifically, data was generated for 400 subjects.
The number of observation times for the response was Poisson distributed
with intensity rate 5, and similarly for the number of observation times
for the covariates. Observation times are generated from a uniform
distribution Unif(0,1) independently. The covariate process is
Gaussian, with values at fixed time points being multivariate normal
with mean 0, variance 1 and correlation \eqn{\exp(-|t_{ij} - t_{ik}|)}{exp(-|t_ij - t_ik|)}.
The responses were generated from \eqn{Y(t) = \beta_0 + X(t)*\beta_1 + \epsilon(t)}{Y(t) = beta0 + X(t)*beta1 + epsilon(t)},
where \eqn{\beta_0}{beta0} = 0.5, \eqn{\beta_1}{beta1} = 0.4t + 0.5, and \eqn{\epsilon(t)}{epsilon(t)} is Gaussian with
mean 0, variance 1 and \eqn{\mathrm{cov}(\epsilon(s),\epsilon(t)) = 2^{-|t-s|}}{cov(e(s),e(t)) = 2^(-|t-s|)}.
Covariates are stored as TD.x. Responses are stored as
TD.y.
}
\format{
TD.x is a data frame with 4052 observations on the following 3 variables.
\describe{
\item{\code{ID}}{patient identifier, there are 400 patients.}
\item{\code{t}}{the covariate observation times}
\item{\code{X1}}{the covariate measured at observation time t}
}
TD.y is a data frame with 3939 observations on the following 3 variables.
\describe{
\item{\code{ID}}{patient identifier, there are 400 patients.}
\item{\code{t}}{the response observation times.}
\item{\code{Y}}{the response measured at time t.}
}
}
\references{
Cao, H., Zeng, D., and Fine, J. P. (2015)
Regression Analysis of sparse asynchronous longitudinal data.
Journal of the Royal Statistical Society: Series B, 77, 755-776.
}
\source{
Generated by Shannon T. Holloway in R.
}
\keyword{datasets }
