\name{AsynchLong-package}
\alias{AsynchLong-package}
\alias{AsynchLong}
\docType{package}
\title{
Regression Analysis of Sparse Asynchronous Longitudinal Data
}
\description{
  Estimation of regression models for sparse asynchronous longitudinal 
  observations, where time-dependent response and covariates are 
  mismatched and observed intermittently within subjects. Kernel 
  weighted estimating equations are
  used for generalized linear models with either time-invariant or 
  time-dependent coefficients.
}
\details{
\tabular{ll}{
  Package: \tab AsynchLong \cr
  Type: \tab Package \cr
  Version: \tab 1.0 \cr
  Date: \tab 2015-02-06 \cr
  License: \tab GPL-2 \cr
}

}
\author{
Hongyuan Cao, Donglin Zeng, Jason P. Fine, and Shannon T. Holloway
Maintainer: Shannon T. Holloway <sthollow@ncsu.edu>
}
\references{
Cao, H., Zeng, D., and Fine, J. P. (2015)
Regression Analysis of sparse asynchronous longitudinal data.
Journal of the Royal Statistical Society: Series B, in press.
}
\keyword{ package }
\seealso{
         \code{\link{asynchTD}}, 
         \code{\link{asynchTI}}, 
         \code{\link{asynchLV}}
}

