\name{AsynchLong-package}
\alias{AsynchLong-package}
\alias{AsynchLong}
\docType{package}
\title{
Regression Analysis of Sparse Asynchronous Longitudinal Data
}
\description{
  Estimation of regression models for sparse asynchronous longitudinal 
  observations, where time-dependent response and covariates are 
  mismatched and observed intermittently within subjects. Kernel 
  weighted estimating equations are
  used for generalized linear models with either time-invariant or 
  time-dependent coefficients.
}
\details{
\tabular{ll}{
  Package: \tab AsynchLong \cr
  Type: \tab Package \cr
  Version: \tab 2.2 \cr
  Date: \tab 2022-06-05 \cr
  License: \tab GPL-2 \cr
}

}
\author{
Hongyuan Cao, Jason P. Fine, Jialiang Li, Donglin Zeng, and Shannon T. Holloway
Maintainer: Shannon T. Holloway <shannon.t.hollowa@gmail.com>
}
\references{
Cao, H., Zeng, D., and Fine, J. P. (2015)
Regression Analysis of sparse asynchronous longitudinal data.
Journal of the Royal Statistical Society: Series B, 77, 755-776.

Cao, H., Li, Jialiang, and Fine, J. P. (2016).
On last observation carried forward and asynchronous longitudinal regression analysis.
Electronic Journal of Statistics, 10, 1155-1180.
}
\keyword{ package }
\seealso{
         \code{\link{asynchTD}}, 
         \code{\link{asynchTI}}, 
         \code{\link{asynchLV}}
}

